applied a large architectural update to how files and folders are organized by adding both a common and pages folders to the components folder.
