+++ src/App.js
@@ -1,14 +1,14 @@
 import React, { useState } from 'react';
 import 'normalize.css';
-import { Home } from './components/Home/Home';
+import { Home } from './components/pages/Home';
 import { Route } from 'react-router-dom';
 import { PrivateRoute } from './components/PrivateRoute/PrivateRoute';
 import GlobalStyles from './styles/GlobalStyles';
 import Typography from './styles/Typography';
 import { PlantContext } from './components/context/PlantContext';
-import Login from './components/Login-Signup/Login';
-import Signup from './components/Login-Signup/Signup';
-import { Dashboard } from './components/Dashboard/Dashboard';
+import Login from './components/pages/Login';
+import Signup from './components/pages/Signup';
+import { Dashboard } from './components/pages/Dashboard';
 
 function App() {
   const [plantId, setPlantId] = useState();
