+++ src/components/common/Notes.js
