+++ src/components/common/Tabs.js
