+++ src/components/common/styles/AddPlant-styling.js
