+++ src/components/common/styles/Calendar-styling.js
