+++ src/components/common/styles/Notes-styling.js
