+++ src/components/common/styles/Tabs-styling.js
