+++ src/components/pages/Signup.js
@@ -2,9 +2,9 @@ import React, { useState, useEffect } from 'react';
 import { Link } from 'react-router-dom';
 import * as yup from 'yup';
 import { axiosWithAuth } from '../../utils/axiosWithAuth';
-import { StyledDiv } from './LoginSignup-styling';
+import { StyledDiv } from './styles/LoginSignup-styling';
 
-export default function Login({ history }) {
+export default function Signup({ history }) {
   const [buttonDisabled, setButtonDisabled] = useState(true);
 
   const [formState, setFormState] = useState({
@@ -32,50 +32,45 @@ export default function Login({ history }) {
     e.preventDefault();
 
     axiosWithAuth()
-      .post('/auth/login', formState)
-      .then((res) => {
-        window.localStorage.clear();
-        window.localStorage.setItem('token', res.data.token);
-        window.localStorage.setItem('id', res.data.user.id);
+      .post('/auth/register', formState)
+      .then(() => {
+        history.push('/login');
 
         setFormState({
           username: '',
           password: '',
         });
-
-        history.push('/dashboard');
       })
       .catch((err) => {
-        console.log('login error: ', err);
+        console.log('Signup error: ', err);
       });
   };
 
-  const inputChange = (event) => {
-    event.persist();
+  const inputChange = (e) => {
+    e.persist();
 
-    const newFormData = {
+    setFormState({
       ...formState,
-      [event.target.name]: event.target.value,
-    };
+      [e.target.name]: e.target.value,
+    });
 
-    validateChange(event);
-    setFormState(newFormData);
+    validateChange(e);
   };
 
-  const validateChange = (event) => {
+  const validateChange = (e) => {
     yup
-      .reach(formSchema, event.target.name)
-      .validate(event.target.value)
+      .reach(formSchema, e.target.name)
+      .validate(e.target.value)
       .then(() => {
         setErrors({
           ...errors,
-          [event.target.name]: '',
+          [e.target.name]: '',
         });
       })
       .catch((err) => {
         setErrors({
           ...errors,
-          [event.target.name]: err.errors[0],
+          [e.target.name]: err.errors[0],
         });
       });
   };
@@ -84,38 +79,38 @@ export default function Login({ history }) {
     <StyledDiv>
       <section>
         <nav>
-          <Link to='/login' className='signin'>
-            Sign In
+          <Link to='/login'>Sign In</Link>
+          <Link to='/signup' className='signup'>
+            Sign up
           </Link>
-          <Link to='/signup'>Sign up</Link>
         </nav>
 
         <div className='form-main'>
-          <h2>Welcome Back!</h2>
+          <h2>Create An Account!</h2>
           <p>
-            Your plants missed you. Don't forget to check up on how they are
-            doing.
+            Keep forgetting to water your plants? No problem. We have the
+            solution. Sign up today.
           </p>
 
           <form onSubmit={handleSubmit}>
             <input
               type='text'
               name='username'
+              placeholder='Username'
               value={formState.username}
               onChange={inputChange}
-              placeholder='Username'
             />
 
             <input
               type='password'
               name='password'
+              placeholder='Password'
               value={formState.password}
               onChange={inputChange}
-              placeholder='Password'
             />
 
             <button disabled={buttonDisabled} type='submit'>
-              Sign In
+              Create Account
             </button>
           </form>
